tests/submission/Opera/getElementsByClassName/014.htm

<!-- quirks mode -->
<html class="a A">
 <head>
  <title>document.getElementsByClassName(): case-insensitive (quirks mode)</title>
  <script src="../resources/testharness.js"></script>
 </head>
 <body class="a a">
  <div id="log"></div>
  <script>test(function() {
             assert_array_equals(document.getElementsByClassName("A a"), 
                                 [document.documentElement, document.body]);
          })
  </script>
 </body>
</html>
